AIRPORT FLIGHT SYSTEM with C
we have added the input read function, Sort function and output function.
In the input reading function, we printed the flight information we received with scanf from the input file using the printf function.
Our other function, input sorting, is the flight we take from the input file
if the information is first followed by the hour, then the priority number
we have sorted it by plane number.